+++ src/main/java/com/smartFarm/was/domain/model/Category.java
@@ -2,7 +2,7 @@ package com.smartFarm.was.domain.model;
 
 public class Category {
 
-    private Integer categoryId;
+    private long categoryId;
     private String status;
     private String categoryType;
 }
